requirements.txt
asgiref==3.9.1
cffi==2.0.0
cryptography==45.0.7
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
